Rename branch-utils to git-utils for broader scope
- Renamed branch-utils.ts to git-utils.ts
- Renamed branch-utils.test.ts to git-utils.test.ts
- Updated imports in main.ts
- Allows for adding other git-related functionality in the future
